fix CI: EnumInfo flags field placement + 32-bit sign-extend + JSON coverage
Three follow-ups from PR #2739 CI + Copilot review.

1. EnumInfo schema bump broke AOT build on linux/macos clang. The new `flags`
   field landed before `hash`, so existing positional initializers
   `{name, module_name, fields, count, hash}` narrowed `UINT64_C(hash)` into
   `uint32_t flags`. Move `flags` after `hash`, expose it via EnumInfoAnnotation,
   and have aot_cpp.das emit the 6th value (`info.flags`). Old initializers
   would have lost the unsigned flag silently, but the AOT regenerator picks
   up the new emitter on the next build.

2. WalkEnumeration (32-bit) had the same sign-extension trap as the 8/16-bit
   walkers. In json_print.cpp, uint32-backed enums with values > INT32_MAX
   came in as negative int32_t, promoted to negative int64_t, and missed the
   lookup against the positive int64_t-stored field value. In debug_print.h,
   the `uvalue = uint64_t(int32_t)` defense-in-depth cast also sign-extended
   via C++ int promotion rules; uint32_t(value) first fixes it. The 8-bit
   and 16-bit equivalents in debug_print.h were already correct (uint8_t and
   uint16_t are smaller than int, so they zero-extend at promotion).

3. Test coverage gaps flagged by Copilot. Add sprint_json assertions for
   uint8/uint16/uint32 enum struct fields in both the language and AOT test
   files — exercises the json_print.cpp WalkEnumeration*/Enum() lookup that
   the first patch reworked. Also adds a uint32 enum (Mask32 with HIGH_BIT
   = 0x80000000u) and asserts its interpolation by name to cover the
   debug_print.h fix.

+++ include/daScript/simulate/debug_print.h
@@ -459,7 +459,10 @@ namespace das {
             }
         }
         virtual void WalkEnumeration ( int32_t & value, EnumInfo * info ) override {
-            int64_t uvalue = uint64_t(value);
+            // uint32_t(value) first, then promote — uint64_t(int32_t) sign-extends via int rules,
+            // which would leave 0x80000000 looking like 0xFFFFFFFF80000000 and miss the uint-backed
+            // enum's stored int64_t value (2147483648).
+            int64_t uvalue = uint64_t(uint32_t(value));
             for ( uint32_t t=0, ts=info->count; t!=ts; ++t ) {
                 if ( value == info->fields[t]->value || uvalue == info->fields[t]->value ) {
                     emitEnumName(info);
